+++ src/Electron/src/Preload/preload.fs
@@ -9,6 +9,7 @@ Remoting.createIpc () |> Remoting.buildTwoWayBridge<IGitApi>
 Remoting.createIpc () |> Remoting.buildTwoWayBridge<IGitLabApi>
 Remoting.createIpc () |> Remoting.buildTwoWayBridge<IAuthApi>
 Remoting.createIpc () |> Remoting.buildTwoWayBridge<ITemplateApi>
+Remoting.createIpc () |> Remoting.buildTwoWayBridge<IValidationPackageIPC>
 
 Remoting.createIpc () |> Remoting.buildBridge<IPathChangeRendererApi>
 Remoting.createIpc () |> Remoting.buildBridge<IRecentArcsRendererApi>